Move Guardian specific API classes to it's own folder
