+++ guardian/src/main/java/com/edoubletech/newsfeed/guardian/api/GuardianFields.kt
@@ -1,0 +1,1 @@
+/* *  Copyright (C) 2018 Eton Otieno Oboch * *  Licensed under the Apache License, Version 2.0 (the "License"); *  you may not use this file except in compliance with the License. *  You may obtain a copy of the License at * *       http://www.apache.org/licenses/LICENSE-2.0 * *  Unless required by applicable law or agreed to in writing, software *  distributed under the License is distributed on an "AS IS" BASIS, *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *  See the License for the specific language governing permissions and *  limitations under the License. * */package com.edoubletech.newsfeed.guardian.apiimport com.google.gson.annotations.Exposeimport com.google.gson.annotations.SerializedNameclass GuardianFields(@SerializedName("thumbnail") @Expose val thumbnail: String,                     @SerializedName("body") @Expose val bodyText: String,                     @SerializedName("trailText") @Expose val trailText: String)
