+++ guardian/src/main/java/com/edoubletech/newsfeed/guardian/api/GuardianResponse.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2018 Eton Otieno Oboch
+ *  Copyright (C) 2019 Eton Otieno
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -12,10 +12,9 @@
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
- *
  */
 
-package com.edoubletech.newsfeed.guardian
+package com.edoubletech.newsfeed.guardian.api
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
